chore(registration-api): Move all configuration classes under configuration package
